+++ app/src/main/AndroidManifest.xml
@@ -13,7 +13,7 @@
         android:theme="@style/Theme.TravelHelper"
         tools:targetApi="31">
         <activity
-            android:name=".screens.flightScreens.FlightActivity"
+            android:name=".screens.flight.FlightActivity"
             android:exported="false" />
         <activity
             android:name=".screens.InitActivity"
