+++ app/src/main/java/ru/itmo/travelhelper/presenter/HotelPresenter.kt
@@ -5,7 +5,7 @@ import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.GlobalScope
 import kotlinx.coroutines.launch
 import ru.itmo.data.repositories.HotelRepositoryImpl
-import ru.itmo.domain.usecases.hotelUseCases.GetHotelsUseCase
+import ru.itmo.domain.usecases.hotel.GetHotelsUseCase
 import ru.itmo.travelhelper.view.HotelView
 
 class HotelPresenter(private val view: HotelView) {
