+++ app/src/main/java/ru/itmo/travelhelper/presenter/flight/FlightDeparturePresenter.kt
@@ -1,0 +1,48 @@
+package ru.itmo.travelhelper.presenter.flight
+
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.GlobalScope
+import kotlinx.coroutines.launch
+import ru.itmo.data.repositories.flight.LocationRepositoryImpl
+import ru.itmo.domain.usecases.flight.GetAirportUseCase
+import ru.itmo.domain.usecases.flight.GetCityUseCase
+import ru.itmo.domain.usecases.flight.GetCountryUseCase
+import ru.itmo.travelhelper.screens.flight.FlightDeparture
+
+class FlightDeparturePresenter(private val view: FlightDeparture) {
+    private val getCountryUseCase = GetCountryUseCase(LocationRepositoryImpl())
+    private val getCityUseCase = GetCityUseCase(LocationRepositoryImpl())
+    private val getAirportUseCase = GetAirportUseCase(LocationRepositoryImpl())
+    private var savedDepartureData = mutableListOf("","","")
+
+    fun setupView() {
+        GlobalScope.launch(Dispatchers.Main) {
+            loadLocationData()
+        }
+    }
+
+
+    private suspend fun loadLocationData() {
+        val countriesData = getCountryUseCase.execute()
+        val citiesData = getCityUseCase.execute()
+        val airportsData = getAirportUseCase.execute()
+
+        view.getCountries(countriesData)
+        view.getCitiesMap(citiesData)
+        view.getAirportMap(airportsData)
+    }
+
+
+    fun updateFullSavedDepartureData(dataToSave: MutableList<String>) {
+        this.savedDepartureData = dataToSave
+    }
+
+    fun updateExactIndexSavedDepartureData(dataIndexToSave: String, index: Int) {
+        this.savedDepartureData[index] = dataIndexToSave
+    }
+
+    fun giveDepartureData(): MutableList<String> {
+        return savedDepartureData
+    }
+
+}
