+++ app/src/main/java/ru/itmo/travelhelper/screens/InitActivity.kt
@@ -3,7 +3,7 @@ package ru.itmo.travelhelper.screens
 import android.content.Intent
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
-import ru.itmo.travelhelper.screens.flightScreens.FlightActivity
+import ru.itmo.travelhelper.screens.flight.FlightActivity
 
 
 class InitActivity : AppCompatActivity() {
