+++ app/src/main/java/ru/itmo/travelhelper/view/flight/FlightDataView.kt
@@ -1,0 +1,6 @@
+package ru.itmo.travelhelper.view.flight
+
+interface FlightDataView {
+
+}
+
