--- app/src/main/java/ru/itmo/travelhelper/view/flightViews/FlightDateFragmentView.kt
@@ -1,6 +1,0 @@
-package ru.itmo.travelhelper.view.flightViews
-
-interface FlightDateFragmentView {
-
-}
-
