+++ app/src/main/res/layout/activity_flight_tickets_main.xml
@@ -5,7 +5,7 @@
     android:id="@+id/main"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".screens.flightScreens.FlightActivity"
+    tools:context=".screens.flight.FlightActivity"
     android:background="@color/background">
 
     <FrameLayout
