+++ app/src/main/res/layout/fragment_flight_location_arrival.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".screens.flightScreens.FlightArrivalFragment">
+    tools:context=".screens.flight.FlightArrivalFragment">
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
