+++ domain/src/main/java/ru/itmo/domain/models/flight/AirportModel.kt
@@ -1,0 +1,5 @@
+package ru.itmo.domain.models.flight
+
+data class AirportModel(
+    val airportName: String,
+)
