+++ domain/src/main/java/ru/itmo/domain/models/flight/CityModel.kt
@@ -1,0 +1,5 @@
+package ru.itmo.domain.models.flight
+
+data class CityModel(
+    val cityName: String,
+)
