+++ domain/src/main/java/ru/itmo/domain/models/flight/CountryModel.kt
@@ -1,0 +1,5 @@
+package ru.itmo.domain.models.flight
+
+data class CountryModel(
+    val countryName: String,
+)
