--- domain/src/main/java/ru/itmo/domain/models/flightModels/AirportModel.kt
@@ -1,5 +1,0 @@
-package ru.itmo.domain.models.flightModels
-
-data class AirportModel(
-    val airportName: String,
-)
