--- domain/src/main/java/ru/itmo/domain/models/flightModels/CityModel.kt
@@ -1,5 +1,0 @@
-package ru.itmo.domain.models.flightModels
-
-data class CityModel(
-    val cityName: String,
-)
