--- domain/src/main/java/ru/itmo/domain/models/flightModels/CountryModel.kt
@@ -1,5 +1,0 @@
-package ru.itmo.domain.models.flightModels
-
-data class CountryModel(
-    val countryName: String,
-)
