+++ domain/src/main/java/ru/itmo/domain/repositories/hotel/HotelRepository.kt
@@ -1,0 +1,9 @@
+package ru.itmo.domain.repositories.hotel
+
+
+import ru.itmo.domain.models.hotelModels.Hotel
+
+interface HotelRepository {
+    suspend fun getHotels(): List<Hotel>
+
+}
