--- domain/src/main/java/ru/itmo/domain/repositories/hotelRepositories/HotelRepository.kt
@@ -1,9 +1,0 @@
-package ru.itmo.domain.repositories.hotelRepositories
-
-
-import ru.itmo.domain.models.hotelModels.Hotel
-
-interface HotelRepository {
-    suspend fun getHotels(): List<Hotel>
-
-}
