+++ domain/src/main/java/ru/itmo/domain/usecases/flight/GetCountryUseCase.kt
@@ -1,0 +1,10 @@
+package ru.itmo.domain.usecases.flight
+
+import ru.itmo.domain.models.flight.CountryModel
+import ru.itmo.domain.repositories.flight.LocationRepository
+
+class GetCountryUseCase(private val locationRepository: LocationRepository) {
+    suspend fun execute(): List<CountryModel> {
+        return locationRepository.getCountries()
+    }
+}
