+++ domain/src/main/java/ru/itmo/domain/usecases/hotel/GetHotelUseCase.kt
@@ -1,0 +1,10 @@
+package ru.itmo.domain.usecases.hotel
+
+import ru.itmo.domain.models.hotelModels.Hotel
+import ru.itmo.domain.repositories.hotel.HotelRepository
+
+class GetHotelsUseCase(private val hotelRepository: HotelRepository) {
+    suspend fun execute(): List<Hotel> {
+        return hotelRepository.getHotels()
+    }
+}
